fix: apply style guide to fix build
