+++ hello-world-do-template/worker-configuration.d.ts
